Movement physics (#1)
* Skit working backwards

* Skidding Works

* SKIDDING COMPLETE

* Efficiency Check

+++ 2DGame/02-Bubble/02-Bubble/TileMap.cpp
@@ -185,7 +185,7 @@ bool TileMap::collisionMoveRight(const glm::ivec2 &pos, const glm::ivec2 &size) 
 	return false;
 }
 
-bool TileMap::collisionMoveDown(const glm::ivec2 &pos, const glm::ivec2 &size, int *posY) const
+bool TileMap::collisionMoveDown(const glm::vec2 &pos, const glm::ivec2 &size, float *posY) const
 {
 	int x0, x1, y;
